refactor: unite socket write with RMS processing - move ResponseBuffer to tcp.rs - make write_fragment it's method - move processing to single task
