+++ src/protocol/rpc/mod.rs
@@ -31,4 +31,4 @@ mod wire;
 
 pub use context::Context;
 pub use transaction_tracker::TransactionTracker;
-pub use wire::{write_fragment, SocketMessageHandler};
+pub use wire::SocketMessageHandler;
